dom/workers/test/serviceworkers/test_empty_serviceworker.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test that registering an empty service worker works</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  function runTest() {
    navigator.serviceWorker.register("empty.js", {scope: "."})
      .then(function(registration) {
        if (registration.installing) {
          registration.installing.onstatechange = function(e) {
            e.target.onstatechange = null;
            done(registration);
          };
        } else {
          done(registration);
        }
      });
  }

  function done(registration) {
    ok(registration.waiting || registration.active, "registration worked");
    registration.unregister().then(function(success) {
      ok(success, "unregister worked");
      SimpleTest.finish();
    });
  }

  SimpleTest.waitForExplicitFinish();
  onload = function() {
    SpecialPowers.pushPrefEnv({"set": [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.enabled", true],
      ["dom.serviceWorkers.testing.enabled", true]
    ]}, runTest);
  };
</script>
</pre>
</body>
</html>
